requirements.txt
fastapi
uvicorn
jinja2
scikit-learn
numpy
python-multipart
requests
python-dotenv
